Remove debug, fix error ret, make norme
